Google Console Aggregated Reports/requirements.txt
google-cloud-storage == 2.11.0
pathlib == 1.0.1
google-cloud-bigquery == 3.12.0
python-dotenv == 1.0.0
pendulum == 3.0.0
